pyproject.toml
[tool.poetry]
name = "lesson-28-dz"
version = "0.1.0"
description = ""
authors = ["Giru <yagofarov.vadim@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
psycopg2 = "^2.9.3"
Pillow = "^9.1.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
